Replace _is_atomic(key) with isinstance(key, Hashable)
Any hashable may be added to a set (OrderedSet). Hashables are sometime
iterables (tuple, frozenset, string). As soon as OrderedSet.index treats
iterables differntly compared to sets, it should be protected against
interating over hashables, which may be in the OrderedSet. Before only
protection for strings and tuples did exist. In this comment a general
approach for any hashable is added. Relevant unit test is added.